requirements.txt
deta==1.1.0
matplotlib==3.5.2
pandas==1.4.2
Pillow==9.1.1
python-dotenv==0.20.0
requests==2.28.0
seaborn==0.11.2
streamlit==1.10.0
streamlit_authenticator==0.1.5
